Fix authorization
In an attempt to not authorize transactions prior to the order being
completed we moved to using a checkout token in some places that don't
work. We were trying to get the transaction on the confirm step but the
original change made it so that there was no transaction. They tried to
get around that by passing in the checkout token but they were still
calling get transaction which would error.

Now we keep track of the checkout token and use that to authorize on
complete.